requirements.txt
Pillow>=9.0.0
reportlab>=4.0.0
numpy>=1.21.0
